backend/requirements.txt
Django==5.2.7
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5,<6
drf-yasg>=1.21,<2
Pillow>=10,<11
django-cors-headers>=4,<5
django-storages[s3]>=1.14,<2
gunicorn>=23,<24
whitenoise>=6.8,<7
psycopg2-binary>=2.9,<3
numpy>=1.26,<3
opencv-python>=4.8,<5
face-recognition>=1.3.0
requests>=2.31,<3
boto3>=1.34,<2
